\name{combinedModuleList}
\alias{combinedModuleList}
\docType{data}
\title{
a list of combined modules
}
\description{
a list of modules combined three resources, CORUM, PCDq, and predicted modules using ClusterONE.
}
\usage{data("combinedModuleList")}
\format{
  The format is:
List of 1118
 $ : chr [1:904] "FBXO4" "FBXL3" "FBXW2" "ISY1" ...
 $ : chr [1:559] "RPL36AL" "L3MBTL2" "MUL1" "RNF166" ...
 $ : chr [1:430] "NOL8" "DHX30" "EIF3M" "MICAL3" ...
 $ : chr [1:315] "MACROD1" "PAAF1" "RIOK1" "LARP6" ...
 $ : chr [1:301] "NPAT" "EIF2B1" "EIF4A2" "TCEB3" ...
 $ : chr [1:296] "DYRK3" "NXNL2" "ANKS6" "NAIF1" ...
 $ : chr [1:287] "TNRC6A" "SH3RF2" "GPX8" "SAP25" ...
 $ : chr [1:253] "PDK3" "MAPRE1" "TSC22D1" "ELF2" ...
 $ : chr [1:214] "SSBP2" "ERH" "RPL19" "SRSF3" ...
 $ : chr [1:196] "NF2" "RDX" "RFC4" "RFC2" ...
 $ : chr [1:183] "TMEM62" "FAAP100" "CPSF1" "KIF1A" ...
 $ : chr [1:172] "TAF1" "TAF10" "TAF11" "TAF2" ...
 $ : chr [1:164] "ACIN1" "AQR" "DDX39B" "BCAS2" ...
 $ : chr [1:144] "ZNF165" "PAPOLA" "BRCA2" "PSMD7" ...
 $ : chr [1:143] "RNH1" "RARG" "ANXA8" "ENO3" ...
 $ : chr [1:136] "BCL6" "HDAC1" "MBD3" "MTA2" ...
 $ : chr [1:133] "TRA2B" "UBE2B" "PPP2R2A" "SNRPN" ...
 $ : chr [1:130] "EIF4A1" "RPS20" "PRPS1" "SHFM1" ...
 $ : chr [1:130] "RPS10" "RPS11" "RPS12" "RPS13" ...
 $ : chr [1:124] "DDX18" "SDAD1" "NLE1" "RBM28" ...
 $ : chr [1:108] "DCAF17" "PSMG4" "FGD3" "RARS2" ...
 $ : chr [1:95] "ARID1A" "ISLR" "BHLHE40" "SBF1" ...
 $ : chr [1:89] "ANXA9" "ACADM" "FGFR1" "INTS9" ...
 $ : chr [1:77] "LMX1A" "TOPBP1" "ZNF516" "DHX38" ...
 $ : chr [1:76] "MRPS10" "MRPS35" "MRPS5" "MRPS36" ...
 $ : chr [1:75] "ALB" "MAP4K3" "TBC1D2" "NUF2" ...
 $ : chr [1:70] "CCT8" "CCT4" "ANXA11" "PAPOLA" ...
 $ : chr [1:70] "SNRNP25" "SNRNP48" "DHX15" "PDCD7" ...
 $ : chr [1:69] "ZNF45" "ZNF33A" "ZNF33B" "PPARA" ...
 $ : chr [1:67] "DAAM2" "PRPF39" "RPAIN" "KTN1" ...
 $ : chr [1:64] "L3HYPDH" "PRICKLE1" "WDR92" "BBC3" ...
 $ : chr [1:63] "ZNF587B" "NACA" "PRPF4B" "HDAC1" ...
 $ : chr [1:62] "PDHA2" "BLVRB" "PRDX5" "PEBP1" ...
 $ : chr [1:62] "GFRA3" "DIAPH1" "TSPAN1" "SULT1C4" ...
 $ : chr [1:61] "RAMP1" "BCAS2" "DNAJC8" "SMNDC1" ...
 $ : chr [1:61] "WDR89" "SDSL" "C5orf30" "MARS2" ...
 $ : chr [1:60] "MAK16" "NIFK" "GTPBP4" "WDR33" ...
 $ : chr [1:60] "USF1" "ME2" "GLDC" "WARS" ...
 $ : chr [1:58] "POM121C" "PODXL" "KIF3C" "NRP1" ...
 $ : chr [1:56] "HNRNPH3" "HNRNPH1" "STIP1" "ARRB2" ...
 $ : chr [1:51] "POLR2F" "PSME3" "MAGOH" "VBP1" ...
 $ : chr [1:51] "GNL1" "LTBR" "GPX4" "HSD17B3" ...
 $ : chr [1:51] "KLHL36" "KIF2B" "EMC1" "KLHDC1" ...
 $ : chr [1:51] "DGKD" "TST" "PCK2" "GALNT6" ...
 $ : chr [1:48] "TSPAN2" "TSPAN3" "SNAP91" "RAMP2" ...
 $ : chr [1:48] "ND1" "ND2" "ND3" "ND4L" ...
 $ : chr [1:46] "PSMC1" "PSMC2" "PSMC3" "PSMC4" ...
 $ : chr [1:45] "RELB" "GALK2" "AMPD2" "CTBS" ...
 $ : chr [1:44] "ZER1" "WWC1" "UEVLD" "LRRC46" ...
 $ : chr [1:44] "MED1" "MED12" "MED14" "MED16" ...
 $ : chr [1:43] "WDR1" "TDO2" "POLD2" "ALDH9A1" ...
 $ : chr [1:43] "IRF6" "KLF11" "PRKAB2" "DHX16" ...
 $ : chr [1:42] "TUBGCP5" "CDK5RAP2" "CEP89" "MAP7D2" ...
 $ : chr [1:42] "EDN1" "LYN" "RET" "ADH4" ...
 $ : chr [1:41] "PSMD12" "PSMD9" "CLIC1" "EIF3F" ...
 $ : chr [1:41] "CYP3A4" "FOXRED1" "C7orf55" "SHD" ...
 $ : chr [1:41] "TMEM110" "VPS36" "JAZF1" "PARG" ...
 $ : chr [1:40] "NUBP1" "RAD23A" "RAD23B" "EIF5" ...
 $ : chr [1:40] "LGALSL" "PARP14" "PLEKHN1" "ZC2HC1C" ...
 $ : chr [1:39] "MUC5AC" "FBLN2" "DGCR2" "NID2" ...
 $ : chr [1:39] "LHX8" "GIPC3" "GIPC2" "CRELD1" ...
 $ : chr [1:38] "BMI1" "COMMD3-BMI1" "WDR11" "CBX3" ...
 $ : chr [1:37] "SSR1" "NOP2" "NSF" "MAP2K3" ...
 $ : chr [1:37] "C11orf49" "TMEM231" "C14orf93" "ZMAT3" ...
 $ : chr [1:37] "DCTN6" "DNM1L" "RTCA" "PIK3C2A" ...
 $ : chr [1:37] "DAGLB" "CHST14" "SPRYD3" "GAL3ST1" ...
 $ : chr [1:37] "CHD2" "GOSR2" "TOR1A" "PRMT5" ...
 $ : chr [1:36] "IKBKB" "CHUK" "MAP3K7" "TUBB4A" ...
 $ : chr [1:35] "ATP5A1" "IVD" "EEF1G" "TKT" ...
 $ : chr [1:34] "STMN1" "YBX3" "ZFX" "RHOQ" ...
 $ : chr [1:34] "LMNB1" "EMD" "LMNB2" "TOMM6" ...
 $ : chr [1:34] "CERK" "RHPN1" "URGCP" "SYTL5" ...
 $ : chr [1:34] "TPP1" "A2M" "GBA" "GAPDH" ...
 $ : chr [1:33] "LTA4H" "HSPA1A" "HSPA1B" "U2AF1L5" ...
 $ : chr [1:33] "EIF4E" "ENO1" "PYGL" "GPI" ...
 $ : chr [1:33] "COA6" "TOR1AIP1" "KRT77" "GALNT5" ...
 $ : chr [1:32] "EXOSC4" "ISYNA1" "NOX4" "DCP1A" ...
 $ : chr [1:32] "ATG4A" "VCPIP1" "NAP1L3" "PIK3R4" ...
 $ : chr [1:32] "XRCC3" "FRS3" "RGS14" "RNF13" ...
 $ : chr [1:31] "TRIM2" "FAM117A" "INO80B" "CCSER2" ...
 $ : chr [1:31] "CNTF" "S100A4" "OTX1" "GPR183" ...
 $ : chr [1:30] "FGF5" "NEFH" "HSD17B1" "GFAP" ...
 $ : chr [1:30] "CHST15" "UBE3D" "PTAR1" "TRAPPC6B" ...
 $ : chr [1:30] "PDE4DIP" "TTI2" "SLC30A6" "DUS1L" ...
 $ : chr [1:30] "BCAS2" "BZW1" "CDC5L" "CWC15" ...
 $ : chr [1:29] "PSMB10" "RBM34" "IL12RB1" "CDKN2B" ...
 $ : chr [1:29] "SMG9" "NAPB" "UPF3A" "WNT8A" ...
 $ : chr [1:29] "AAGAB" "ATF7IP" "PACS1" "RNF111" ...
 $ : chr [1:29] "STEAP3" "MAVS" "GLRX5" "SUPV3L1" ...
 $ : chr [1:29] "CHD2" "GOSR2" "TOR1A" "PRMT5" ...
 $ : chr [1:29] "ERCC2" "ERCC3" "GTF2B" "GTF2E1" ...
 $ : chr [1:28] "ARL4D" "PSEN1" "RGS7" "PSEN2" ...
 $ : chr [1:28] "PTRF" "PKN3" "ERICH5" "KLHL35" ...
 $ : chr [1:28] "RAD17" "RFC2" "RFC3" "RFC4" ...
 $ : chr [1:27] "MAST3" "RASAL1" "SNAPIN" "PAK4" ...
 $ : chr [1:27] "DNAAF1" "ACD" "ZNF501" "LRRC58" ...
 $ : chr [1:27] "EPOR" "TNNI1" "CYP2B6" "CYP3A5" ...
 $ : chr [1:27] "ZNF740" "SIMC1" "TMEM104" "NGDN" ...
 $ : chr [1:27] "HS3ST1" "ERLIN2" "DCAF17" "FIGN" ...
  [list output truncated]
}
\references{
MoonFinder: a framework for the identification of moonlighting non-coding RNAs.
}
\examples{
data(combinedModuleList)
head(combinedModuleList)
}
\keyword{datasets}
