\name{corumModuleList}
\alias{corumModuleList}
\docType{data}
\title{
modules from CORUM
}
\description{
a list of modules collected from database CORUM.
}
\usage{data("corumModuleList")}
\format{
  The format is:
List of 464
 $ corum_1  : chr [1:5] "NCAPD2" "NCAPG" "NCAPH" "SMC2" ...
 $ corum_2  : chr [1:6] "CORO2A" "GPS2" "HDAC3" "NCOR1" ...
 $ corum_3  : chr [1:8] "BLOC1S1" "BLOC1S2" "BLOC1S3" "BLOC1S4" ...
 $ corum_4  : chr [1:7] "ACTR2" "ACTR3" "ARPC1B" "ARPC2" ...
 $ corum_5  : chr [1:20] "PSMC1" "PSMC2" "PSMC3" "PSMC4" ...
 $ corum_6  : chr [1:6] "PFDN1" "PFDN2" "PFDN4" "PFDN5" ...
 $ corum_7  : chr [1:8] "AP1B1" "AP1G1" "AP1G2" "AP1M1" ...
 $ corum_8  : chr [1:5] "BCL6" "HDAC1" "MBD3" "MTA2" ...
 $ corum_9  : chr [1:7] "DNMT3B" "HDAC1" "KIF4A" "SIN3A" ...
 $ corum_10 : chr [1:7] "HDAC1" "HDAC2" "RBBP4" "RBBP7" ...
 $ corum_11 : chr [1:5] "GPS2" "HDAC3" "NCOR2" "TBL1X" ...
 $ corum_12 : chr [1:7] "AP3B1" "AP3B2" "AP3D1" "AP3M1" ...
 $ corum_13 : chr [1:7] "CHD4" "HDAC1" "HDAC2" "MBD3" ...
 $ corum_14 : chr [1:8] "CHD4" "HDAC1" "HDAC2" "MBD2" ...
 $ corum_15 : chr [1:8] "MED1" "MED12" "MED14" "MED16" ...
 $ corum_16 : chr [1:10] "ACTB" "ARID1A" "CARM1" "SCYL1" ...
 $ corum_17 : chr [1:9] "NUP107" "NUP133" "NUP160" "NUP37" ...
 $ corum_18 : chr [1:8] "EXOC1" "EXOC2" "EXOC3" "EXOC4" ...
 $ corum_19 : chr [1:8] "ANAPC1" "ANAPC2" "ANAPC4" "ANAPC5" ...
 $ corum_20 : chr [1:9] "ANAPC1" "ANAPC10" "ANAPC2" "ANAPC4" ...
 $ corum_21 : chr [1:24] "ERCC2" "ERCC3" "GTF2B" "GTF2E1" ...
 $ corum_22 : chr [1:12] "POLR2A" "POLR2B" "POLR2C" "POLR2D" ...
 $ corum_23 : chr [1:5] "EED" "EZH2" "RBBP4" "RBBP7" ...
 $ corum_24 : chr [1:9] "CCNH" "CDK7" "ERCC2" "ERCC3" ...
 $ corum_25 : chr [1:5] "SNAPC1" "SNAPC2" "SNAPC3" "SNAPC4" ...
 $ corum_26 : chr [1:12] "BMI1" "COMMD3-BMI1" "CBX2" "CBX4" ...
 $ corum_27 : chr [1:13] "BMI1" "COMMD3-BMI1" "CBX2" "CBX4" ...
 $ corum_28 : chr [1:8] "CD81" "ADGRG1" "GNA11" "GNAQ" ...
 $ corum_29 : chr [1:10] "POP1" "POP4" "POP5" "POP7" ...
 $ corum_30 : chr [1:8] "TCP1" "CCT2" "CCT3" "CCT4" ...
 $ corum_31 : chr [1:6] "ACD" "POT1" "TERF1" "TERF2" ...
 $ corum_32 : chr [1:5] "APH1A" "BSG" "NCSTN" "PSEN1" ...
 $ corum_33 : chr [1:9] "AP1B1" "AP1G1" "AP1G2" "AP1M1" ...
 $ corum_34 : chr [1:12] "ARID2" "PBRM1" "SMARCA4" "SMARCB1" ...
 $ corum_35 : chr [1:9] "TAF1" "TAF10" "TAF11" "TAF2" ...
 $ corum_36 : chr [1:7] "NCAPD2" "NCAPG" "NCAPH" "PARP1" ...
 $ corum_37 : chr [1:5] "NCAPD3" "NCAPG2" "NCAPH2" "SMC2" ...
 $ corum_38 : chr [1:8] "COG1" "COG2" "COG3" "COG4" ...
 $ corum_39 : chr [1:44] "ND1" "ND2" "ND3" "ND4L" ...
 $ corum_40 : chr [1:22] "PSMA1" "PSMA2" "PSMA3" "PSMA4" ...
 $ corum_41 : chr [1:5] "ABI1" "BRK1" "CYFIP1" "NCKAP1" ...
 $ corum_42 : chr [1:13] "SMARCA2" "SMARCA4" "SMARCB1" "SMARCC1" ...
 $ corum_43 : chr [1:14] "PSMA1" "PSMA2" "PSMA3" "PSMA4" ...
 $ corum_44 : chr [1:16] "PSMA1" "PSMA2" "PSMA3" "PSMA4" ...
 $ corum_45 : chr [1:36] "PSMA1" "PSMA2" "PSMA3" "PSMA4" ...
 $ corum_46 : chr [1:15] "PSMA1" "PSMA2" "PSMA3" "PSMA4" ...
 $ corum_47 : chr [1:5] "CUL2" "RBX1" "ELOC" "ELOB" ...
 $ corum_48 : chr [1:5] "ASB2" "CUL5" "RNF7" "ELOC" ...
 $ corum_49 : chr [1:5] "ASB1" "CUL5" "RNF7" "ELOC" ...
 $ corum_50 : chr [1:5] "ASB6" "CUL5" "RNF7" "ELOC" ...
 $ corum_51 : chr [1:5] "ASB7" "CUL5" "RNF7" "ELOC" ...
 $ corum_52 : chr [1:5] "ASB12" "CUL5" "RNF7" "ELOC" ...
 $ corum_53 : chr [1:5] "CUL5" "RBX1" "ELOC" "ELOB" ...
 $ corum_54 : chr [1:8] "MED1" "MED14" "MED17" "MED23" ...
 $ corum_55 : chr [1:8] "CCNC" "CDK8" "MED1" "MED14" ...
 $ corum_56 : chr [1:7] "CCNC" "CDK8" "MED10" "MED14" ...
 $ corum_57 : chr [1:32] "CCNC" "CDK19" "CDK8" "MED1" ...
 $ corum_58 : chr [1:15] "ACAD8" "MED1" "MED12" "MED13" ...
 $ corum_59 : chr [1:11] "ACTL6A" "BRCA1" "SMARCA2" "SMARCA4" ...
 $ corum_60 : chr [1:13] "BLM" "FANCA" "FANCC" "FANCE" ...
 $ corum_61 : chr [1:9] "FAAP100" "FANCA" "FANCB" "FANCC" ...
 $ corum_62 : chr [1:6] "CCNB1" "EPN1" "NUMB" "RALBP1" ...
 $ corum_63 : chr [1:11] "BLM" "FANCA" "FANCC" "FANCE" ...
 $ corum_64 : chr [1:5] "RAD51C" "RAD51B" "RAD51D" "XRCC2" ...
 $ corum_65 : chr [1:5] "RAD17" "RFC2" "RFC3" "RFC4" ...
 $ corum_66 : chr [1:8] "HUS1" "RAD1" "RAD17" "RAD9A" ...
 $ corum_67 : chr [1:5] "RFC1" "RFC2" "RFC3" "RFC4" ...
 $ corum_68 : chr [1:5] "GAPDH" "HMGB1" "HMGB2" "HSPA8" ...
 $ corum_69 : chr [1:16] "BAZ1A" "CHD3" "HDAC1" "HDAC2" ...
 $ corum_70 : chr [1:11] "MED1" "MED14" "MED15" "MED17" ...
 $ corum_71 : chr [1:14] "CCNC" "CDK8" "MED1" "MED12" ...
 $ corum_72 : chr [1:15] "CCNC" "CDK8" "MED1" "MED12" ...
 $ corum_73 : chr [1:5] "PCNA" "POLD1" "POLD2" "POLD3" ...
 $ corum_74 : chr [1:6] "APEX1" "HIF1A" "STAT3" "CITED1" ...
 $ corum_75 : chr [1:12] "MED10" "MED14" "MED16" "MED17" ...
 $ corum_76 : chr [1:17] "CCNC" "CDK8" "MED1" "MED10" ...
 $ corum_77 : chr [1:13] "ACTL6A" "ACTR5" "ACTR8" "INO80C" ...
 $ corum_78 : chr [1:10] "ACTL6A" "ACTR6" "EAF2" "H2AFZ" ...
 $ corum_79 : chr [1:34] "RPS10" "RPS11" "RPS12" "RPS13" ...
 $ corum_80 : chr [1:81] "RPL10" "RPL10A" "RPL11" "RPL12" ...
 $ corum_81 : chr [1:47] "RPL10" "RPL10A" "RPL11" "RPL12" ...
 $ corum_82 : chr [1:6] "CCNB1" "CCNB2" "CCND1" "CDK1" ...
 $ corum_83 : chr [1:5] "CCND1" "CCND3" "CDK5" "CDKN1A" ...
 $ corum_84 : chr [1:30] "DAP3" "MRPS10" "MRPS11" "MRPS12" ...
 $ corum_85 : chr [1:76] "DAP3" "LACTB" "MRPL1" "MRPL10" ...
 $ corum_86 : chr [1:46] "LACTB" "MRPL1" "MRPL10" "MRPL11" ...
 $ corum_87 : chr [1:31] "RPS10" "RPS11" "RPS12" "RPS13" ...
 $ corum_88 : chr [1:5] "SGCA" "SGCB" "SGCD" "SGCG" ...
 $ corum_89 : chr [1:144] "ACIN1" "AQR" "DDX39B" "BCAS2" ...
 $ corum_90 : chr [1:6] "MLH1" "MSH2" "MSH6" "PCNA" ...
 $ corum_91 : chr [1:6] "MCM2" "MCM3" "MCM4" "MCM5" ...
 $ corum_92 : chr [1:6] "ORC1" "ORC2" "ORC3" "ORC4" ...
 $ corum_93 : chr [1:24] "SNRNP25" "SNRNP48" "DHX15" "PDCD7" ...
 $ corum_94 : chr [1:12] "ATM" "BLM" "BRCA1" "MLH1" ...
 $ corum_95 : chr [1:8] "BLM" "BRCA1" "MLH1" "MSH2" ...
 $ corum_96 : chr [1:6] "ATM" "BRCA1" "MLH1" "RFC1" ...
 $ corum_97 : chr [1:10] "BRCA1" "KAT2A" "MSH2" "MSH6" ...
 $ corum_98 : chr [1:11] "ESR1" "KAT2A" "MED1" "MED12" ...
 $ corum_99 : chr [1:8] "BACH1" "ERCC3" "GJB6" "KLF10" ...
  [list output truncated]
}
\references{
MoonFinder: a framework for the identification of moonlighting non-coding RNAs.
}
\examples{
data(corumModuleList)
head(corumModuleList)
}
\keyword{datasets}
