\name{moduleList}
\alias{moduleList}
\docType{data}
\title{
ncRNA target moudles
}
\description{
a list of moudles targeted by ncRNAs, which is a subset of the combinedModuleList. Because not all the moduels are targeted by a ncRNA.
}
\usage{data("moduleList")}
\format{
  The format is:
List of 175
 $ mapped_1  : chr [1:172] "TAF1" "TAF10" "TAF11" "TAF2" ...
 $ mapped_2  : chr [1:164] "ACIN1" "AQR" "DDX39B" "BCAS2" ...
 $ mapped_3  : chr [1:136] "BCL6" "HDAC1" "MBD3" "MTA2" ...
 $ mapped_4  : chr [1:130] "RPS10" "RPS11" "RPS12" "RPS13" ...
 $ mapped_5  : chr [1:76] "DAP3" "MRPS10" "MRPS11" "MRPS12" ...
 $ mapped_6  : chr [1:70] "SNRNP25" "SNRNP48" "DHX15" "PDCD7" ...
 $ mapped_7  : chr [1:48] "ND1" "ND2" "ND3" "ND4L" ...
 $ mapped_8  : chr [1:46] "PSMC1" "PSMC2" "PSMC3" "PSMC4" ...
 $ mapped_9  : chr [1:44] "MED1" "MED12" "MED14" "MED16" ...
 $ mapped_10 : chr [1:38] "BMI1" "COMMD3-BMI1" "WDR11" "CBX3" ...
 $ mapped_11 : chr [1:36] "IKBKB" "CHUK" "MAP3K7" "TUBB4A" ...
 $ mapped_12 : chr [1:34] "TPP1" "A2M" "GBA" "GAPDH" ...
 $ mapped_13 : chr [1:30] "BCAS2" "BZW1" "CDC5L" "CWC15" ...
 $ mapped_14 : chr [1:29] "ERCC2" "ERCC3" "GTF2B" "GTF2E1" ...
 $ mapped_15 : chr [1:28] "RAD17" "RFC2" "RFC3" "RFC4" ...
 $ mapped_16 : chr [1:27] "ASH2L" "C17orf49" "INO80C" "CHD8" ...
 $ mapped_17 : chr [1:26] "ACTL6A" "ACTR6" "EAF2" "H2AFZ" ...
 $ mapped_18 : chr [1:25] "PCNA" "POLD1" "POLD2" "POLD3" ...
 $ mapped_19 : chr [1:25] "CD3EAP" "CHUK" "CUL1" "FBXW11" ...
 $ mapped_20 : chr [1:24] "ACD" "POT1" "TERF1" "TERF2" ...
 $ mapped_21 : chr [1:24] "ASH2L" "DPY30" "MEN1" "POLR2B" ...
 $ mapped_22 : chr [1:23] "ACTB" "API5" "C1QBP" "CKB" ...
 $ mapped_23 : chr [1:22] "NFKB1" "NFKB2" "REL" "RELA" ...
 $ mapped_24 : chr [1:21] "PIK3R2" "ABL1" "EGFR" "EGF" ...
 $ mapped_25 : chr [1:20] "CUL2" "RBX1" "ELOC" "ELOB" ...
 $ mapped_26 : chr [1:20] "DDX1" "DDX17" "DDX3X" "DDX5" ...
 $ mapped_27 : chr [1:20] "BTRC" "CUL1" "FBXW11" "NFKBIA" ...
 $ mapped_28 : chr [1:19] "BLM" "FANCA" "FANCC" "FANCE" ...
 $ mapped_29 : chr [1:19] "CDC37" "HSP90AA1" "HSP90AB1" "HSPA4" ...
 $ mapped_30 : chr [1:19] "CDC37" "HSP90AA1" "HSP90AB1" "IKBKB" ...
 $ mapped_31 : chr [1:19] "ITGB3" "ITGB1" "ITGAV" "ITGA5" ...
 $ mapped_32 : chr [1:18] "BLOC1S1" "BLOC1S2" "BLOC1S3" "BLOC1S4" ...
 $ mapped_33 : chr [1:18] "ABCF2" "CDC5L" "DDB1" "DDX23" ...
 $ mapped_34 : chr [1:18] "DDX39A" "TUBB4A" "TUBB" "TUBB4B" ...
 $ mapped_35 : chr [1:17] "ORC1" "ORC2" "ORC3" "ORC4" ...
 $ mapped_36 : chr [1:17] "COPS2" "COPS3" "COPS4" "COPS5" ...
 $ mapped_37 : chr [1:17] "CAPRIN1" "CLP1" "CPSF3" "CPSF6" ...
 $ mapped_38 : chr [1:17] "INTS11" "INTS1" "INTS10" "INTS12" ...
 $ mapped_39 : chr [1:16] "EIF3A" "EIF3B" "EIF3C" "EIF3D" ...
 $ mapped_40 : chr [1:16] "DIS3" "EXOSC1" "EXOSC2" "EXOSC3" ...
 $ mapped_41 : chr [1:16] "CUL4A" "CUL4B" "DDB1" "DDB2" ...
 $ mapped_42 : chr [1:16] "ACTB" "BANF1" "BCCIP" "CDC37" ...
 $ mapped_43 : chr [1:16] "FOS" "JUN" "FOSL1" "FOSL2" ...
 $ mapped_44 : chr [1:16] "POLR1C" "CD3EAP" "POLR1A" "POLR2E" ...
 $ mapped_45 : chr [1:15] "ATP5A1" "ATP5B" "ATP5C1" "ATP5D" ...
 $ mapped_46 : chr [1:15] "CCNA2" "CDK2" "LIG1" "PARP1" ...
 $ mapped_47 : chr [1:14] "BMI1" "COMMD3-BMI1" "CBX2" "CBX4" ...
 $ mapped_48 : chr [1:14] "CENPC" "CENPH" "CENPM" "CENPN" ...
 $ mapped_49 : chr [1:14] "ATM" "FANCD2" "MDC1" "MRE11" ...
 $ mapped_50 : chr [1:14] "CDK7" "CREBBP" "EP300" "POLR2A" ...
 $ mapped_51 : chr [1:14] "CDC37" "CHUK" "FBL" "HSP90AA1" ...
 $ mapped_52 : chr [1:14] "LIN37" "LIN52" "LIN54" "LIN9" ...
 $ mapped_53 : chr [1:14] "PPP2R1A" "PPP2R1B" "PPP2CB" "PPP2R2A" ...
 $ mapped_54 : chr [1:14] "GAGE7" "GAGE12I" "GAGE12G" "GAGE12F" ...
 $ mapped_55 : chr [1:13] "CCNH" "CDK7" "ERCC2" "ERCC3" ...
 $ mapped_56 : chr [1:13] "ACTL6A" "ACTR5" "ACTR8" "INO80C" ...
 $ mapped_57 : chr [1:13] "CHMP1A" "CHMP1B" "CHMP2A" "CHMP2B" ...
 $ mapped_58 : chr [1:13] "SF3B1" "CD2BP2" "U2AF2" "YWHAG" ...
 $ mapped_59 : chr [1:12] "ACTR2" "ACTR3" "ARPC1B" "ARPC2" ...
 $ mapped_60 : chr [1:12] "E2F6" "EHMT1" "EHMT2" "GJB3" ...
 $ mapped_61 : chr [1:12] "SEPT11" "SEPT2" "SEPT7" "SEPT8" ...
 $ mapped_62 : chr [1:12] "CHUK" "ELP1" "IKBKB" "MAP3K14" ...
 $ mapped_63 : chr [1:12] "MEAF6" "ING4" "ING5" "KAT7" ...
 $ mapped_64 : chr [1:12] "ANKRD28" "BTRC" "CHUK" "CUL1" ...
 $ mapped_65 : chr [1:12] "PSMD11" "SMAD7" "SMAD9" "SMAD6" ...
 $ mapped_66 : chr [1:12] "CASP1" "CASP14" "CASP3" "CASP5" ...
 $ mapped_67 : chr [1:11] "CORO2A" "GPS2" "HDAC3" "NCOR1" ...
 $ mapped_68 : chr [1:11] "MCM2" "MCM3" "MCM4" "MCM5" ...
 $ mapped_69 : chr [1:11] "CASC3" "EIF4A3" "MAGOH" "RBM8A" ...
 $ mapped_70 : chr [1:11] "DDX21" "DHX9" "HNRNPC" "PRPF8" ...
 $ mapped_71 : chr [1:11] "GTF3C1" "GTF3C2" "GTF3C3" "GTF3C4" ...
 $ mapped_72 : chr [1:11] "CPSF1" "CPSF2" "CPSF3" "CPSF4" ...
 $ mapped_73 : chr [1:11] "DDX20" "DHX9" "GEMIN4" "GTF2B" ...
 $ mapped_74 : chr [1:11] "BCAR1" "CBL" "CRK" "GRB2" ...
 $ mapped_75 : chr [1:11] "DARS" "EEF1E1" "EPRS" "IARS" ...
 $ mapped_76 : chr [1:11] "DDX17" "DDX5" "DGCR8" "DHX9" ...
 $ mapped_77 : chr [1:11] "HDAC1" "HIPK2" "RBBP4" "SIN3A" ...
 $ mapped_78 : chr [1:11] "ACTN1" "BCAR1" "CDH1" "CTNNB1" ...
 $ mapped_79 : chr [1:11] "NCK1" "SFPQ" "PDIA3" "LRPAP1" ...
 $ mapped_80 : chr [1:11] "TNF" "TNFRSF1A" "TNFRSF1B" "RPS3" ...
 $ mapped_81 : chr [1:11] "IRAK2" "IL1R1" "IRAK1" "TIRAP" ...
 $ mapped_82 : chr [1:11] "TAPBP" "HLA-B" "HLA-A" "HLA-E" ...
 $ mapped_83 : chr [1:11] "NR5A2" "EDF1" "RARA" "RXRA" ...
 $ mapped_84 : chr [1:10] "AP1B1" "AP1G1" "AP1G2" "AP1M1" ...
 $ mapped_85 : chr [1:10] "ANAPC1" "ANAPC2" "ANAPC4" "ANAPC5" ...
 $ mapped_86 : chr [1:10] "POP1" "POP4" "POP5" "POP7" ...
 $ mapped_87 : chr [1:10] "COG1" "COG2" "COG3" "COG4" ...
 $ mapped_88 : chr [1:10] "AKAP9" "KCNQ1" "PPP1CA" "PPP1CB" ...
 $ mapped_89 : chr [1:10] "LSM1" "LSM2" "LSM3" "LSM4" ...
 $ mapped_90 : chr [1:10] "ACTB" "ACTN4" "HSP90AB1" "HSPA8" ...
 $ mapped_91 : chr [1:10] "DCP2" "EXOSC10" "EXOSC2" "EXOSC4" ...
 $ mapped_92 : chr [1:10] "KMT2A" "KMT2D" "KAT8" "RBBP5" ...
 $ mapped_93 : chr [1:10] "HDAC3" "NCOR1" "SF3A1" "SF3B3" ...
 $ mapped_94 : chr [1:10] "CTCF" "H2AFZ" "HIST2H2AA3" "HIST2H2AA4" ...
 $ mapped_95 : chr [1:10] "ACTB" "HSPA4" "MYH10" "NCL" ...
 $ mapped_96 : chr [1:10] "CDCA5" "PDS5A" "PDS5B" "RAD21" ...
 $ mapped_97 : chr [1:10] "BACH1" "BARD1" "BRCA1" "MRE11" ...
 $ mapped_98 : chr [1:10] "ARID1B" "CREBBP" "NCOA3" "SMAD2" ...
 $ mapped_99 : chr [1:10] "AZI2" "CDC37" "HSP90AA1" "HSP90AB1" ...
  [list output truncated]
}
\references{
MoonFinder: a framework for the identification of moonlighting non-coding RNAs.
}
\examples{
data(moduleList)
head(moduleList)
}
\keyword{datasets}
