\name{pcdqModuleList}
\alias{pcdqModuleList}
\docType{data}
\title{
pcdqModuleList
}
\description{
a list of modules collected from database PCDq.
}
\usage{data("pcdqModuleList")}
\format{
  The format is:
List of 412
 $ pcdq_1  : chr [1:39] "TPP1" "A2M" "GBA" "GAPDH" ...
 $ pcdq_2  : chr [1:41] "IKBKB" "CHUK" "MAP3K7" "TUBB4A" ...
 $ pcdq_3  : chr [1:24] "MED7" "MED7" "MED14" "MED6" ...
 $ pcdq_4  : chr [1:12] "BRCA1" "MYC" "ESR1" "TP53" ...
 $ pcdq_5  : chr [1:23] "PIK3R2" "ABL1" "EGFR" "EGF" ...
 $ pcdq_6  : chr [1:19] "FOS" "FOS" "JUN" "FOSL1" ...
 $ pcdq_7  : chr [1:14] "SF3B1" "CD2BP2" "U2AF2" "YWHAG" ...
 $ pcdq_8  : chr [1:10] "KAT7" "VIM" "SAT1" "CHD3" ...
 $ pcdq_9  : chr [1:6] "RTCA" "HLA-B" "IMPDH2" "MCM3" ...
 $ pcdq_10 : chr [1:18] "ZNHIT1" "YEATS4" "ACTL6A" "MORF4L2" ...
 $ pcdq_11 : chr [1:23] "PSMA7" "EIF1B" "EIF1B" "EPB41" ...
 $ pcdq_12 : chr [1:14] "PSMD11" "SMAD7" "SMAD9" "SMAD6" ...
 $ pcdq_13 : chr [1:13] "NCK1" "SFPQ" "SFPQ" "PDIA3" ...
 $ pcdq_14 : chr [1:17] "ORC5" "ORC5" "ORC4" "CDK1" ...
 $ pcdq_15 : chr [1:7] "APOA1" "TTR" "ALB" "TF" ...
 $ pcdq_16 : chr [1:7] "HGS" "STAM2" "DAZAP2" "UBQLN4" ...
 $ pcdq_17 : chr [1:6] "LDOC1" "HNRNPC" "FXR2" "CCDC85B" ...
 $ pcdq_18 : chr [1:7] "PLSCR1" "SPRY2" "GLRX3" "EFEMP2" ...
 $ pcdq_19 : chr [1:7] "BCL2" "BID" "BAX" "BCL2L1" ...
 $ pcdq_20 : chr [1:7] "GH1" "RPL3" "RPL3" "EIF6" ...
 $ pcdq_21 : chr [1:11] "CFLAR" "FAS" "FASLG" "FASLG" ...
 $ pcdq_22 : chr [1:10] "PIAS1" "TP53" "AR" "SUMO1" ...
 $ pcdq_23 : chr [1:12] "RB1" "MYBL1" "MYBL2" "RBL2" ...
 $ pcdq_24 : chr [1:8] "FYB1" "FYN" "HCK" "SRC" ...
 $ pcdq_25 : chr [1:12] "TNF" "TNF" "TNFRSF1A" "TNFRSF1B" ...
 $ pcdq_26 : chr [1:8] "FAM107A" "FAM107A" "KRT19" "KRT15" ...
 $ pcdq_27 : chr [1:10] "RBFOX2" "CCNK" "ATXN1" "DAZAP2" ...
 $ pcdq_28 : chr [1:13] "WDR5" "RBX1" "TLE2" "CUL4A" ...
 $ pcdq_29 : chr [1:24] "PSMD11" "PSMD12" "PSMD14" "PSMD3" ...
 $ pcdq_30 : chr [1:5] "SPTBN1" "SPTAN1" "CEP63" "EXOC1" ...
 $ pcdq_31 : chr [1:20] "GEMIN2" "SNRPB" "SNRPB" "GEMIN4" ...
 $ pcdq_32 : chr [1:11] "ORC4" "MCM3" "MCM4" "MCM5" ...
 $ pcdq_33 : chr [1:16] "DXO" "ZFP36" "EXOSC10" "EXOSC9" ...
 $ pcdq_34 : chr [1:6] "CTNND1" "CDH1" "CTNNA1" "CTNNB1" ...
 $ pcdq_35 : chr [1:11] "RAD21" "MYC" "SMC1A" "PDS5A" ...
 $ pcdq_36 : chr [1:9] "CDC42" "RAC1" "YWHAH" "PRKCZ" ...
 $ pcdq_37 : chr [1:6] "RFXANK" "HNRNPM" "C20orf24" "NUP93" ...
 $ pcdq_38 : chr [1:7] "MCC" "VHL" "ELOC" "CUL2" ...
 $ pcdq_39 : chr [1:14] "POLR2D" "POLR2C" "POLR2C" "POLR2E" ...
 $ pcdq_40 : chr [1:11] "BRCA1" "NBN" "CHEK2" "H2AFX" ...
 $ pcdq_41 : chr [1:8] "NUP214" "NUP62" "NUP153" "RANBP2" ...
 $ pcdq_42 : chr [1:8] "IFNAR1" "TYK2" "IL6ST" "IL6ST" ...
 $ pcdq_43 : chr [1:18] "ARID1A" "ACTL6A" "SMARCA2" "SMARCA4" ...
 $ pcdq_44 : chr [1:64] "PRMT5" "RAD51B" "HIST1H2BK" "HIST1H2BJ" ...
 $ pcdq_45 : chr [1:9] "HRAS" "RAF1" "BRAF" "MAPK1" ...
 $ pcdq_46 : chr [1:7] "TUBG1" "AKT1" "AKT2" "GSK3A" ...
 $ pcdq_47 : chr [1:8] "THRB" "VDR" "RXRA" "PRMT2" ...
 $ pcdq_48 : chr [1:8] "MEN1" "RPA2" "RPA1" "RPA3" ...
 $ pcdq_49 : chr [1:8] "APAF1" "CASP1" "CASP3" "CASP9" ...
 $ pcdq_50 : chr [1:9] "BACH1" "CREB3" "CREB1" "CREB1" ...
 $ pcdq_51 : chr [1:7] "TXNDC9" "PTN" "CSTF2" "BAG6" ...
 $ pcdq_52 : chr [1:20] "SAP18" "HDAC3" "NCOR1" "SAP30" ...
 $ pcdq_53 : chr [1:5] "GFI1B" "ACTN1" "TRIM27" "MAGEA11" ...
 $ pcdq_54 : chr [1:13] "ARPC1B" "ARPC2" "ARPC2" "ARPC3" ...
 $ pcdq_55 : chr [1:7] "KPNA3" "ARRB2" "ARRB2" "CSNK2A1" ...
 $ pcdq_56 : chr [1:7] "TP53" "HTT" "NAP1L1" "CTBP1" ...
 $ pcdq_57 : chr [1:9] "EIF4G3" "EIF4E" "EIF4E" "EIF4A2" ...
 $ pcdq_58 : chr [1:10] "SNRNP200" "MYC" "HNRNPC" "TOP1" ...
 $ pcdq_59 : chr [1:9] "FANCG" "FANCG" "FANCA" "FANCC" ...
 $ pcdq_60 : chr [1:7] "DVL2" "ZNF250" "FXR2" "AP2M1" ...
 $ pcdq_61 : chr [1:6] "SNAPIN" "BLOC1S1" "BLOC1S4" "BLOC1S2" ...
 $ pcdq_62 : chr [1:8] "DDX6" "ZFP36" "DCP1A" "EDC4" ...
 $ pcdq_63 : chr [1:7] "GDF9" "ZNF24" "APLP1" "EEF1A1" ...
 $ pcdq_64 : chr [1:13] "IRAK2" "IL1R1" "IRAK1" "TIRAP" ...
 $ pcdq_65 : chr [1:10] "COPS2" "GPS1" "COPS4" "COPS3" ...
 $ pcdq_66 : chr [1:6] "SMC2" "NCAPH" "NCAPD2" "NEK6" ...
 $ pcdq_67 : chr [1:7] "WWP2" "EWSR1" "HERPUD1" "HERPUD1" ...
 $ pcdq_68 : chr [1:14] "PPP2R1A" "PPP2R1B" "PPP2CB" "PPP2R2A" ...
 $ pcdq_69 : chr [1:5] "GAS7" "NCKAP1" "ABI1" "CYFIP2" ...
 $ pcdq_70 : chr [1:9] "CBL" "CBLB" "PDCD6IP" "SH3KBP1" ...
 $ pcdq_71 : chr [1:9] "CDK4" "CCND1" "CCND1" "CCND3" ...
 $ pcdq_72 : chr [1:11] "ACTL6A" "TCF3" "RUVBL2" "INO80C" ...
 $ pcdq_73 : chr [1:10] "LSM1" "LSM2" "LSM8" "SNRPE" ...
 $ pcdq_74 : chr [1:25] "SF3B1" "SNRPB" "SNRPB" "U2AF2" ...
 $ pcdq_75 : chr [1:5] "BCR" "CFTR" "COPB1" "PDZK1" ...
 $ pcdq_76 : chr [1:7] "TMSB4X" "LIMS1" "ERH" "ILK" ...
 $ pcdq_77 : chr [1:6] "GNAI2" "FLNA" "CANX" "MTNR1A" ...
 $ pcdq_78 : chr [1:6] "TPD52L2" "TPD52L2" "TPD52" "TPD52L1" ...
 $ pcdq_79 : chr [1:9] "KRT6A" "KRT6C" "KRT18" "KRT8" ...
 $ pcdq_80 : chr [1:7] "ERCC2" "ERCC3" "ERCC5" "GTF2H1" ...
 $ pcdq_81 : chr [1:5] "DNMT1" "DNMT1" "SUZ12" "EZH2" ...
 $ pcdq_82 : chr [1:9] "TGFB1" "TGFB3" "ENG" "ENG" ...
 $ pcdq_83 : chr [1:6] "CRYAA" "CRYAA2" "CRYAB" "HSPB1" ...
 $ pcdq_84 : chr [1:6] "MAFG" "MAFK" "JUN" "NFE2L2" ...
 $ pcdq_85 : chr [1:20] "ITGB3" "ITGB1" "ITGAV" "ITGA5" ...
 $ pcdq_86 : chr [1:8] "PARP1" "XRCC6" "XRCC5" "XRCC1" ...
 $ pcdq_87 : chr [1:7] "CDC37" "CCT7" "TANK" "TANK" ...
 $ pcdq_88 : chr [1:9] "CBX4" "E2F6" "BMI1" "COMMD3-BMI1" ...
 $ pcdq_89 : chr [1:8] "CDK11B" "CDK11A" "YWHAQ" "YWHAB" ...
 $ pcdq_90 : chr [1:6] "PEX14" "ABCD3" "ABCD1" "PEX19" ...
 $ pcdq_91 : chr [1:10] "POP7" "RPP14" "POP4" "RPP38" ...
 $ pcdq_92 : chr [1:7] "UBE2M" "NEDD8" "CAV1" "CAV1" ...
 $ pcdq_93 : chr [1:6] "F2" "VTN" "ITGB3" "SERPINE1" ...
 $ pcdq_94 : chr [1:5] "CD4" "NR3C1" "LCK" "HSP90AA1" ...
 $ pcdq_95 : chr [1:10] "NDC80" "ZWINT" "CBX5" "CBX3" ...
 $ pcdq_96 : chr [1:6] "ASB1" "ASB6" "ASB12" "CUL5" ...
 $ pcdq_97 : chr [1:7] "MYC" "PCNA" "RFC4" "RFC2" ...
 $ pcdq_98 : chr [1:7] "MEN1" "WDR5" "KMT2A" "RBBP5" ...
 $ pcdq_99 : chr [1:6] "TOM1L1" "PTBP1" "HNRNPH1" "CRMP1" ...
  [list output truncated]
}
\references{
MoonFinder: a framework for the identification of moonlighting non-coding RNAs.
}
\examples{
data(pcdqModuleList)
head(pcdqModuleList)
}
\keyword{datasets}
