\name{predModuleList}
\alias{predModuleList}
\docType{data}
\title{
predicted modules
}
\description{
a list of protein moudles identified using ClusterONE.
}
\usage{data("predModuleList")}
\format{
  The format is:
List of 837
 $ pred_1  : chr [1:216] "FBXO4" "FBXL3" "FBXW2" "ISY1" ...
 $ pred_2  : chr [1:169] "TNRC6A" "SH3RF2" "GPX8" "SAP25" ...
 $ pred_3  : chr [1:17] "NXT1" "NUDT5" "IKZF2" "FBXO5" ...
 $ pred_4  : chr [1:228] "PDK3" "MAPRE1" "TSC22D1" "ELF2" ...
 $ pred_5  : chr [1:51] "POLR2F" "PSME3" "MAGOH" "VBP1" ...
 $ pred_6  : chr [1:17] "FAU" "RPL23" "UBE2D2" "RPS15" ...
 $ pred_7  : chr [1:559] "RPL36AL" "L3MBTL2" "MUL1" "RNF166" ...
 $ pred_8  : chr [1:22] "NUBP1" "RAD23A" "RAD23B" "EIF5" ...
 $ pred_9  : chr [1:60] "MAK16" "NIFK" "GTPBP4" "WDR33" ...
 $ pred_10 : chr [1:296] "DYRK3" "NXNL2" "ANKS6" "NAIF1" ...
 $ pred_11 : chr [1:10] "EHBP1" "GPKOW" "SNAPC3" "CLP1" ...
 $ pred_12 : chr [1:144] "ZNF165" "PAPOLA" "BRCA2" "PSMD7" ...
 $ pred_13 : chr [1:46] "DAAM2" "PRPF39" "RPAIN" "KTN1" ...
 $ pred_14 : chr [1:863] "FANCL" "RBFOX1" "BABAM1" "IRAK4" ...
 $ pred_15 : chr [1:80] "TRA2B" "UBE2B" "PPP2R2A" "SNRPN" ...
 $ pred_16 : chr [1:77] "LMX1A" "TOPBP1" "ZNF516" "DHX38" ...
 $ pred_17 : chr [1:63] "ZNF587B" "NACA" "PRPF4B" "HDAC1" ...
 $ pred_18 : chr [1:253] "ELP5" "LMX1A" "SH3RF2" "GPX8" ...
 $ pred_19 : chr [1:45] "RELB" "GALK2" "AMPD2" "CTBS" ...
 $ pred_20 : chr [1:6] "SMG9" "NAPB" "UPF3A" "WNT8A" ...
 $ pred_21 : chr [1:69] "ZNF45" "ZNF33A" "ZNF33B" "PPARA" ...
 $ pred_22 : chr [1:430] "NOL8" "DHX30" "EIF3M" "MICAL3" ...
 $ pred_23 : chr [1:186] "NPAT" "EIF2B1" "EIF4A2" "TCEB3" ...
 $ pred_24 : chr [1:315] "MACROD1" "PAAF1" "RIOK1" "LARP6" ...
 $ pred_25 : chr [1:6] "DDX18" "SDAD1" "NLE1" "RBM28" ...
 $ pred_26 : chr [1:56] "HNRNPH3" "HNRNPH1" "STIP1" "ARRB2" ...
 $ pred_27 : chr [1:43] "WDR1" "TDO2" "POLD2" "ALDH9A1" ...
 $ pred_28 : chr [1:14] "SLC25A16" "GLB1" "PPP3CB" "PFKL" ...
 $ pred_29 : chr [1:32] "EXOSC4" "ISYNA1" "NOX4" "DCP1A" ...
 $ pred_30 : chr [1:296] "NACA" "PRPF4B" "HDAC1" "TDG" ...
 $ pred_31 : chr [1:199] "TSC22D1" "NAB2" "TAB1" "MAPK11" ...
 $ pred_32 : chr [1:33] "LTA4H" "HSPA1A" "HSPA1B" "U2AF1L5" ...
 $ pred_33 : chr [1:6] "RBMX" "IFNGR2" "EIF4A3" "IGHMBP2" ...
 $ pred_34 : chr [1:37] "CCT8" "CCT4" "ANXA11" "PAPOLA" ...
 $ pred_35 : chr [1:12] "GLTSCR1L" "ZMAT3" "CWC22" "XAB2" ...
 $ pred_36 : chr [1:183] "TMEM62" "FAAP100" "CPSF1" "KIF1A" ...
 $ pred_37 : chr [1:99] "HBG2" "PSPH" "RBM6" "RPP30" ...
 $ pred_38 : chr [1:214] "SSBP2" "ERH" "RPL19" "SRSF3" ...
 $ pred_39 : chr [1:60] "USF1" "ME2" "GLDC" "WARS" ...
 $ pred_40 : chr [1:17] "NAT10" "DDX47" "ESF1" "HEATR1" ...
 $ pred_41 : chr [1:134] "RNH1" "RARG" "ANXA8" "ENO3" ...
 $ pred_42 : chr [1:13] "ZNF251" "ZNF2" "ZNF426" "ZNF416" ...
 $ pred_43 : chr [1:25] "METTL25" "DDX51" "KRI1" "SRFBP1" ...
 $ pred_44 : chr [1:21] "EIF4A1" "RPS20" "PRPS1" "SHFM1" ...
 $ pred_45 : chr [1:23] "SMC4" "MAN2C1" "SACM1L" "OLA1" ...
 $ pred_46 : chr [1:55] "BRPF1" "EIF6" "SIK1" "WDR4" ...
 $ pred_47 : chr [1:62] "PDHA2" "BLVRB" "PRDX5" "PEBP1" ...
 $ pred_48 : chr [1:24] "CNGB1" "GAPVD1" "HERPUD1" "LAPTM4A" ...
 $ pred_49 : chr [1:31] "TRIM2" "FAM117A" "INO80B" "CCSER2" ...
 $ pred_50 : chr [1:7] "RBM22" "CWC25" "CWC15" "CPSF2" ...
 $ pred_51 : chr [1:6] "ZNF160" "VTA1" "ZNF331" "CELSR1" ...
 $ pred_52 : chr [1:8] "ZNF324B" "ZNF470" "ZNF425" "ZNF529" ...
 $ pred_53 : chr [1:9] "ZNF225" "ZNF221" "ZNF510" "ZNF256" ...
 $ pred_54 : chr [1:65] "ISYNA1" "NOX4" "RIC8A" "RRAGD" ...
 $ pred_55 : chr [1:37] "SSR1" "NOP2" "NSF" "MAP2K3" ...
 $ pred_56 : chr [1:29] "PSMB10" "RBM34" "IL12RB1" "CDKN2B" ...
 $ pred_57 : chr [1:58] "POM121C" "PODXL" "KIF3C" "NRP1" ...
 $ pred_58 : chr [1:26] "WDR61" "INTS2" "NELFA" "TAF1D" ...
 $ pred_59 : chr [1:20] "GNL1" "LTBR" "GPX4" "HSD17B3" ...
 $ pred_60 : chr [1:14] "RFC4" "RFC2" "RFC1" "GTF2F1" ...
 $ pred_61 : chr [1:14] "U2AF1L4" "PPIL4" "CCDC12" "LEO1" ...
 $ pred_62 : chr [1:8] "MRPS28" "MRPS7" "MRPS2" "COQ4" ...
 $ pred_63 : chr [1:25] "SUPT20H" "ARMC2" "NCOA7" "SLC30A5" ...
 $ pred_64 : chr [1:26] "MAPK3" "MARK3" "RPA1" "APEX1" ...
 $ pred_65 : chr [1:27] "MAST3" "RASAL1" "SNAPIN" "PAK4" ...
 $ pred_66 : chr [1:7] "STBD1" "FEN1" "MLH1" "IL15" ...
 $ pred_67 : chr [1:34] "STMN1" "YBX3" "ZFX" "RHOQ" ...
 $ pred_68 : chr [1:23] "RINT1" "BZW1" "NOXA1" "LAPTM4B" ...
 $ pred_69 : chr [1:61] "RAMP1" "BCAS2" "DNAJC8" "SMNDC1" ...
 $ pred_70 : chr [1:121] "MLXIPL" "ISYNA1" "NOX4" "DCP1A" ...
 $ pred_71 : chr [1:49] "MAN1A1" "SPR" "CBS" "GCK" ...
 $ pred_72 : chr [1:14] "ZNF721" "ZFP90" "ZNF473" "ZNF480" ...
 $ pred_73 : chr [1:16] "ZNF433" "ZNF283" "CNTD1" "ZNF614" ...
 $ pred_74 : chr [1:69] "RBM25" "EIF2B2" "HINT1" "STK19" ...
 $ pred_75 : chr [1:24] "SULT1C4" "EPRS" "HSP90AA1" "GALT" ...
 $ pred_76 : chr [1:8] "AHI1" "HUS1B" "ALMS1" "CCHCR1" ...
 $ pred_77 : chr [1:75] "ALB" "MAP4K3" "TBC1D2" "NUF2" ...
 $ pred_78 : chr [1:37] "C11orf49" "TMEM231" "C14orf93" "ZMAT3" ...
 $ pred_79 : chr [1:21] "CFB" "LHB" "PRL" "GCG" ...
 $ pred_80 : chr [1:19] "GIPC3" "GIPC2" "BTBD6" "LRR1" ...
 $ pred_81 : chr [1:12] "ZNF667" "ZNF484" "ZNF569" "ZNF684" ...
 $ pred_82 : chr [1:23] "MRPS5" "MRPS36" "MRPS11" "MRPS15" ...
 $ pred_83 : chr [1:14] "DDX47" "ESF1" "HEATR1" "PIF1" ...
 $ pred_84 : chr [1:12] "C14orf93" "FBRS" "GPBP1L1" "ZNF532" ...
 $ pred_85 : chr [1:11] "ZNF605" "ZNF554" "ZNF429" "ZNF677" ...
 $ pred_86 : chr [1:9] "ZNF251" "ZNF2" "ZNF426" "EFCAB11" ...
 $ pred_87 : chr [1:12] "RMI1" "CLSPN" "FANCE" "UBE2T" ...
 $ pred_88 : chr [1:40] "LGALSL" "PARP14" "PLEKHN1" "ZC2HC1C" ...
 $ pred_89 : chr [1:48] "TSPAN2" "TSPAN3" "SNAP91" "RAMP2" ...
 $ pred_90 : chr [1:8] "NUP37" "NUP35" "CASC5" "FBXO30" ...
 $ pred_91 : chr [1:12] "APOL2" "LBHD1" "ACBD6" "ZFP14" ...
 $ pred_92 : chr [1:29] "SMG9" "NAPB" "UPF3A" "WNT8A" ...
 $ pred_93 : chr [1:33] "EIF4E" "ENO1" "PYGL" "GPI" ...
 $ pred_94 : chr [1:11] "XDH" "RFX3" "PIP4K2A" "SOX2" ...
 $ pred_95 : chr [1:26] "COX20" "RBM48" "CELF3" "SZT2" ...
 $ pred_96 : chr [1:14] "ZNF613" "ZNF790" "ZNF562" "ZNF782" ...
 $ pred_97 : chr [1:8] "SNAPC4" "CPSF3L" "RPRD2" "TAF3" ...
 $ pred_98 : chr [1:24] "ZNF821" "DAB1" "PLCG2" "KLRC2" ...
 $ pred_99 : chr [1:15] "RANGAP1" "NUP153" "RANBP2" "GSK3A" ...
  [list output truncated]
}
\references{
MoonFinder: a framework for the identification of moonlighting non-coding RNAs.
}
\examples{
data(predModuleList)
head(predModuleList)
}
\keyword{datasets}
