\name{uniGene}
\alias{uniGene}
\docType{data}
\title{
Background gene symbols
}
\description{
All the gene symbols used in the entire protien interaction network.
}
\usage{data("uniGene")}
\format{
  The format is:
 chr [1:10111] "A1CF" "AAAS" "AADAT" "AAGAB" "AAK1" "AAMDC" "AAMP" "AAR2" "AARS" "AARS2" "AARSD1" "AASDHPPT" ...
}
\references{
MoonFinder: a framework for the identification of moonlighting non-coding RNAs.
}
\examples{
data(uniGene)
head(uniGene)
}
\keyword{datasets}
